+++ test/tsd/index.test-d.ts
@@ -1,6 +1,0 @@
-import middleware from '../../'
-
-middleware({dataFormat: function(data, isRes, req) {
-    req.url === '';
-    return JSON.stringify(data);
-}});